SourceTreeScript/requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
gitdb2==2.0.0
GitPython==2.1.1
Markdown==2.6.7
pyperclip==1.5.27
requests==2.11.1
smmap2==2.0.1
